+++ src/components/fix/FixAdminDetail.tsx
@@ -7,9 +7,10 @@ import RightArrow from '@/assets/rightArrow.svg';
 import { useAdminFixInfo } from '@/hooks/api/fixzone/useAdminFixInfo';
 import { useUpdateComplete } from '@/hooks/api/fixzone/useUpdateComplete';
 import useModal from '@/hooks/common/useModal';
-import { FixAdminDetailType } from '@/types/fixzone';
+import { FixAdminDetailType } from '@/types/fix';
 import { parseImgUrl } from '@/utils/parse';
 import FixItemInfo from './FixItemInfo';
+import Heading from '../common/Heading';
 import Modal from '../common/Modal';
 import ConfirmModal from '../modal/ConfirmModal';
 type Prop = {
@@ -39,43 +40,30 @@ export default function FixAdminDetail({ id }: Prop) {
   const { club, content, createdAt, imageUrls, completed, location, title } =
     data;
   const mutation = useUpdateComplete();
-
+  console.log('data', data);
   function handleCompleted() {
     setData((prev) => ({ ...prev, completed: true }));
     mutation.mutate({ id, completed: true, token });
   }
 
   return (
-    <div className="w-full bg-gray-100">
-      <div className="m-auto max-w-[650px] bg-gray-100 p-10 ">
-        <div className="flex justify-between">
-          <Link href="/fixzone">
-            <Image src={LeftArrow} alt="back" width={25} height={25} />
-          </Link>
-          <div className="text-lg font-bold">동아리방 시설 보수</div>
-          <div></div>
-        </div>
-        <div className="flex justify-end">
-          <button
-            disabled={completed}
-            onClick={openModal}
-            className={`mb-3 mt-7 rounded-xl border border-gray-300 px-4 py-2  text-gray-500  ${
-              !completed
-                ? ` hover:border-green-300 hover:text-green-500`
-                : `border-green-300 text-green-500`
-            } md:mr-0.5`}
-          >
-            {completed ? `처리 완료` : `처리 마치기`}
-          </button>
-        </div>
+    <>
+      <Heading>동아리방 시설보수 확인</Heading>
+      <div className="mt-14 flex items-center">
+        <Link href="/fix">
+          <Image src={LeftArrow} alt="back" width={25} height={25} />
+        </Link>
+        <span className="text-xl font-semibold text-gray-600">{title}</span>
+      </div>
+
+      <div className="mt-3 flex w-full flex-col rounded-xl border border-gray-100 p-6 md:mt-7 md:flex-row">
         {/* 정보 */}
-        <div className="mb-7 rounded-xl bg-white p-5 text-gray-500 shadow-xl">
-          <div className="border-b py-2 text-xl font-bold ">{title}</div>
-          <div className="py-2 pt-4 font-semibold">{content}</div>
+        <div className=" w-1/2 rounded-xl bg-white p-3">
+          <FixItemInfo club={club} createdAt={createdAt} location={location} />
+          <div className="mt-4 py-2 pt-4">{content}</div>
         </div>
-        <FixItemInfo club={club} createdAt={createdAt} location={location} />
         {/* 내용 */}
-        <div className="relative my-7 flex items-center justify-center">
+        <div className="relative flex w-full items-center justify-center p-3 md:w-1/2">
           <Image
             src={LeftArrow}
             width={30}
@@ -94,7 +82,7 @@ export default function FixAdminDetail({ id }: Prop) {
             height={500}
             priority
             alt="fixImage"
-            className="h-[60vh] w-full overflow-hidden object-scale-down "
+            className="overflow-hidden object-scale-down "
           />
           <Image
             src={RightArrow}
@@ -109,19 +97,32 @@ export default function FixAdminDetail({ id }: Prop) {
             }`}
           />
         </div>
-        <Modal
-          visible={visible}
-          modalRef={modalRef}
-          title={'처리하시겠습니까?'}
-          closeModal={closeModal}
-        >
-          <ConfirmModal
-            title=""
-            callback={handleCompleted}
-            closeModal={closeModal}
-          />
-        </Modal>
       </div>
-    </div>
+      <div>
+        <button
+          disabled={completed}
+          onClick={openModal}
+          className={`mb-3 mt-7 rounded-xl border border-gray-300 px-4 py-2  text-gray-500  ${
+            !completed
+              ? ` hover:border-green-300 hover:text-green-500`
+              : `border-green-300 text-green-500`
+          } md:mr-0.5`}
+        >
+          {completed ? `처리 완료` : `처리 마치기`}
+        </button>
+      </div>
+      <Modal
+        visible={visible}
+        modalRef={modalRef}
+        title={'처리하시겠습니까?'}
+        closeModal={closeModal}
+      >
+        <ConfirmModal
+          title=""
+          callback={handleCompleted}
+          closeModal={closeModal}
+        />
+      </Modal>
+    </>
   );
 }
