+++ src/components/fix/FixItemInfo.tsx
@@ -1,0 +1,33 @@
+import Image from 'next/image';
+import Canlender from '@/assets/calender.svg';
+import Group from '@/assets/group.svg';
+import Pin from '@/assets/pin.svg';
+type Props = {
+  club: string;
+  createdAt: string;
+  location: string;
+};
+export default function FixItemInfo({ club, createdAt, location }: Props) {
+  return (
+    <div className="flex justify-between rounded-xl bg-white font-semibold  text-gray-500 ">
+      <div className="flex flex-col items-center justify-center ">
+        <div className="p-3 text-xl font-bold leading-none text-blue-500">
+          동아리
+        </div>
+        <div className=" text-base leading-none"> {club.toUpperCase()}</div>
+      </div>
+      <div className="flex flex-col items-center justify-center">
+        <div className="p-3 text-xl font-bold leading-none text-blue-500">
+          신청일
+        </div>
+        <div className="text-base font-semibold leading-none"> {createdAt}</div>
+      </div>
+      <div className="flex flex-col items-center justify-center">
+        <div className="p-3 text-xl font-bold leading-none text-blue-500">
+          동아리방
+        </div>
+        <div className="text-base font-semibold leading-none"> {location}</div>
+      </div>
+    </div>
+  );
+}
