+++ src/pages/admin/fix/index.tsx
@@ -23,7 +23,7 @@ export default function Index() {
       </Head>
       <div>
         <div className=" flex items-center justify-between text-2xl font-bold leading-tight md:flex md:text-3xl">
-          <Heading>동아리방 시설보수 신청</Heading>
+          <Heading>동아리방 시설보수 확인</Heading>
           {role === ROLE_TYPE.ROLE_CLUB && (
             <Link href="/fix/new">
               <button className=" mt-7 rounded-lg bg-blue-100 px-5 py-2 text-base font-bold text-blue-500 transition-colors hover:bg-blue-200 md:w-auto">
